+++ pydantic_br/field_erros.py
@@ -1,7 +1,4 @@
-try:
-    from pydantic import PydanticTypeError, PydanticValueError
-except ModuleNotFoundError:
-    raise ModuleNotFoundError("Are you sure you installed pydantic")
+from .field_class_errors import get_pydantic_errors_class
 
 __all__ = [
     "FieldTypeError",
@@ -11,21 +8,28 @@ __all__ = [
 ]
 
 
+PydanticTypeError, PydanticValueError = get_pydantic_errors_class()
+
+
 class FieldTypeError(PydanticTypeError):
     code = "not_str"
-    msg_template = "str type expected"
+    msg_template = "Input should be a valid string"
+    message_template = msg_template
 
 
 class FieldMaskError(PydanticValueError):
     code = "invalid_mask"
     msg_template = "invalid mask format"
+    message_template = msg_template
 
 
 class FieldDigitError(PydanticValueError):
     code = "not_digits"
     msg_template = "field only accept digits as string"
+    message_template = msg_template
 
 
 class FieldInvalidError(PydanticValueError):
     code = "invalid_data"
     msg_template = "invalid data"
+    message_template = msg_template